+++ lib/anycable/health_server.rb
@@ -4,42 +4,63 @@ require 'webrick'
 require 'anycable/server'
 
 module Anycable
-  # Server for HTTP healthchecks
-  module HealthServer
-    class << self
-      def start(port)
-        return if running?
+  # Server for HTTP healthchecks.
+  #
+  # Basic usage:
+  #
+  #  # create a new healthcheck server for a specified
+  #  # gRPC server lisening on the port
+  #  health_server = Anycable::HealthServer.new(grpc_server, port)
+  #
+  #  # start health server in background
+  #  health_server.start
+  #
+  #  # stop health server
+  #  health_server.stop
+  class HealthServer
+    SUCCESS_RESPONSE = [200, "Ready"].freeze
+    FAILURE_RESPONSE = [503, "Not Ready"].freeze
 
-        @health_server ||= build_server(port)
-        Thread.new { @health_server.start }
+    attr_reader :grpc_server, :port, :path, :server
 
-        Anycable.logger.info "HTTP health server is listening on #{port}"
-      end
+    def initialize(grpc_server, port:, path: "/health", logger: Anycable.logger)
+      @grpc_server = grpc_server
+      @port = port
+      @path = path
+      @logger = logger
+      @server = build_server
+    end
 
-      def stop
-        return unless running?
+    def start
+      return if running?
 
-        @health_server.shutdown
-      end
+      Thread.new { server.start }
 
-      def running?
-        @health_server&.status == :Running
-      end
+      logger.info "HTTP health server is listening on #{port}"
+    end
 
-      private
+    def stop
+      return unless running?
 
-      SUCCESS_RESPONSE = [200, "Ready"].freeze
-      FAILURE_RESPONSE = [503, "Not Ready"].freeze
+      server.shutdown
+    end
 
-      def build_server(port)
-        WEBrick::HTTPServer.new(
-          Port: port,
-          Logger: Anycable.logger,
-          AccessLog: []
-        ).tap do |server|
-          server.mount_proc '/health' do |_, res|
-            res.status, res.body = Anycable::Server.running? ? SUCCESS_RESPONSE : FAILURE_RESPONSE
-          end
+    def running?
+      server.status == :Running
+    end
+
+    private
+
+    attr_reader :logger
+
+    def build_server
+      WEBrick::HTTPServer.new(
+        Port: port,
+        Logger: logger,
+        AccessLog: []
+      ).tap do |server|
+        server.mount_proc path do |_, res|
+          res.status, res.body = grpc_server.running? ? SUCCESS_RESPONSE : FAILURE_RESPONSE
         end
       end
     end